refactor(types): normalize types-ambient.js
