+++ packages/zoe/src/contracts/callSpread/calculateShares.js
@@ -1,4 +1,4 @@
-/// <reference path="./types.js" />
+/// <reference path="./types-ambient.js" />
 
 import { AmountMath, isNatValue } from '@agoric/ertp';
 import { assert } from '@agoric/assert';
