+++ packages/zoe/src/contracts/exported.js
@@ -1,4 +1,4 @@
-/// <reference path="./types.js" />
+/// <reference path="./types-ambient.js" />
 // eslint-disable-next-line no-restricted-syntax
-import './loan/types.js';
-import './callSpread/types.js';
+import './loan/types-ambient.js';
+import './callSpread/types-ambient.js';
